+++ horus_deploy/deploys.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2021 Horus View and Explore B.V.
+# Copyright (C) 2021-2022 Horus View and Explore B.V.
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -18,7 +18,9 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+import enum
 from pathlib import Path
+from typing import List
 
 from ._config import user_config_dir
 from .metadata import extract_metadata
@@ -28,7 +30,13 @@ _BUILTIN = Path(__file__).parent / "builtin_deploy_scripts"
 _USER_DIR = user_config_dir() / "deploy_scripts"
 
 
-def find_deploy_scripts(deploy_scripts=None):
+class Type(enum.IntFlag):
+    BUILTIN = enum.auto()
+    USER = enum.auto()
+    LOCAL = enum.auto()
+
+
+def find_deploy_scripts(filter_by: List = None):
     """
     Find and load metadata of deploy scripts.
 
@@ -37,25 +45,28 @@ def find_deploy_scripts(deploy_scripts=None):
 
     A list of dictionaries is returned. Each dictionary contains the
     script's metadata (name, description, paramaters), id, and path.
+
+    ``filter_by`` limits the search in directories with a list of names.
     """
-    search_directories = [Path.cwd(), _USER_DIR, _BUILTIN]
+    search_dirs = [Path.cwd(), _USER_DIR, _BUILTIN]
     excludes = ["__init__.py"]
     scripts = []
     search_paths = []
 
-    if not deploy_scripts:
-        for dir in search_directories:
-            if dir.exists():
-                search_paths.extend(dir.iterdir())
+    # Build scearch paths.
+    if filter_by:
+        for f in filter_by:
+            search_paths.extend([(d / f) for d in search_dirs])
     else:
-        for ds in deploy_scripts:
-            for dir in search_directories:
-                search_paths.append(dir / ds)
+        for d in search_dirs:
+            if d.exists():
+                search_paths.extend(d.iterdir())
 
+    # Search for scripts in paths and extract metadata.
     for fp in search_paths:
         if fp.name in excludes:
             continue
-        fp = _fix_path(fp)
+        fp = _expand_path(fp)
         if not fp.exists():
             continue
 
@@ -65,13 +76,16 @@ def find_deploy_scripts(deploy_scripts=None):
 
         metadata["id"] = fp.parent.stem if fp.name == "deploy.py" else fp.stem
         metadata["path"] = fp
+        metadata["type"] = _get_type(fp)
 
         scripts.append(metadata)
+
+    scripts.sort(key=lambda e: (e["type"], e["id"]))
 
     return scripts
 
 
-def _fix_path(p):
+def _expand_path(p):
     if p.is_dir():
         p = p / "deploy.py"
     elif p.suffix != ".py":
@@ -87,17 +101,14 @@ def _get_metadata(fp):
     return metadata
 
 
-def shorten_deploy_script_path(path):
-    path = path.resolve()
-
-    if path.parent == Path.cwd() or path.parent.parent == Path.cwd():
-        if path.name == "deploy.py":
-            path = path.parent
-        path = f"./{path.relative_to(Path.cwd())}"
-    elif is_relative_to(path, Path.home()):
-        path = f"~/{path.relative_to(Path.home())}"
-
-    return path
+def _get_type(path):
+    if is_relative_to(path, _BUILTIN):
+        return Type.BUILTIN
+    if is_relative_to(path, _USER_DIR):
+        return Type.USER
+    if is_relative_to(path, Path.cwd()):
+        return Type.LOCAL
+    raise ValueError("cannot infer type from {path=}")
 
 
 def is_relative_to(p, *other):